decouple APM and client_helper
