+++ src/mongocxx/test/spec/monitoring.hh
@@ -27,6 +27,10 @@ using namespace mongocxx;
 
 // Stores and compares apm events.
 class apm_checker {
+    using assert_matches = std::function<void(bsoncxx::types::bson_value::view,
+                                              bsoncxx::types::bson_value::view,
+                                              mongocxx::spec::apm_checker&)>;
+
    public:
     enum class event { kill_cursors, get_more, configure_fail_point };
     static event to_event(stdx::string_view s);
@@ -39,7 +43,9 @@ class apm_checker {
                  bool allow_extra = false,
                  const test_util::match_visitor& match_visitor = {});
 
-    void compare_v2(bsoncxx::array::view expected, bool allow_extra = false);
+    void compare_v2(bsoncxx::array::view expectations,
+                    assert_matches matches_fn,
+                    bool allow_extra = false);
 
     // Check that the apm checker has all expected events, ignore ordering and extra events.
     void has(bsoncxx::array::view expected);
